Refactor modules from `core` into `xarray.computation`
