+++ xarray/computation/ops.py
@@ -1,6 +1,6 @@
 """Define core operations for xarray objects.
 
-TODO(shoyer): rewrite this module, making use of xarray.core.computation,
+TODO(shoyer): rewrite this module, making use of xarray.computation.computation,
 NumPy's __array_ufunc__ and mixin classes instead of the unintuitive "inject"
 functions.
 """
@@ -8,6 +8,7 @@ functions.
 from __future__ import annotations
 
 import operator
+from typing import Literal
 
 import numpy as np
 
@@ -143,7 +144,7 @@ def fillna(data, other, join="left", dataset_join="left"):
         - "left": take only variables from the first object
         - "right": take only variables from the last object
     """
-    from xarray.core.computation import apply_ufunc
+    from xarray.computation.computation import apply_ufunc
 
     return apply_ufunc(
         duck_array_ops.fillna,
@@ -172,10 +173,10 @@ def where_method(self, cond, other=dtypes.NA):
     -------
     Same type as caller.
     """
-    from xarray.core.computation import apply_ufunc
+    from xarray.computation.computation import apply_ufunc
 
     # alignment for three arguments is complicated, so don't support it yet
-    join = "inner" if other is dtypes.NA else "exact"
+    join: Literal["inner", "exact"] = "inner" if other is dtypes.NA else "exact"
     return apply_ufunc(
         duck_array_ops.where_method,
         self,
