+++ xarray/computation/weighted.py
@@ -6,9 +6,9 @@ from typing import TYPE_CHECKING, Generic, Literal, cast
 import numpy as np
 from numpy.typing import ArrayLike
 
+from xarray.computation.computation import apply_ufunc, dot
 from xarray.core import duck_array_ops, utils
 from xarray.core.alignment import align, broadcast
-from xarray.core.computation import apply_ufunc, dot
 from xarray.core.types import Dims, T_DataArray, T_Xarray
 from xarray.namedarray.utils import is_duck_dask_array
