+++ xarray/core/duck_array_ops.py
@@ -467,8 +467,6 @@ def _ignore_warnings_if(condition):
 
 
 def _create_nan_agg_method(name, coerce_strings=False, invariant_0d=False):
-    from xarray.core import nanops
-
     def f(values, axis=None, skipna=None, **kwargs):
         if kwargs.pop("out", None) is not None:
             raise TypeError(f"`out` is not valid for {name}")
@@ -495,6 +493,8 @@ def _create_nan_agg_method(name, coerce_strings=False, invariant_0d=False):
                 or dtypes.is_object(values.dtype)
             )
         ):
+            from xarray.computation import nanops
+
             nanname = "nan" + name
             func = getattr(nanops, nanname)
         else:
