+++ xarray/core/missing.py
@@ -12,9 +12,9 @@ from typing import TYPE_CHECKING, Any, TypeVar, get_args
 import numpy as np
 import pandas as pd
 
+from xarray.computation.computation import apply_ufunc
 from xarray.core import utils
 from xarray.core.common import _contains_datetime_like_objects, ones_like
-from xarray.core.computation import apply_ufunc
 from xarray.core.duck_array_ops import (
     datetime_to_numeric,
     push,
