+++ xarray/tests/test_computation.py
@@ -10,8 +10,7 @@ import pytest
 from numpy.testing import assert_allclose, assert_array_equal
 
 import xarray as xr
-from xarray.core.alignment import broadcast
-from xarray.core.computation import (
+from xarray.computation.computation import (
     _UFuncSignature,
     apply_ufunc,
     broadcast_compat_data,
@@ -22,6 +21,7 @@ from xarray.core.computation import (
     result_name,
     unified_dim_sizes,
 )
+from xarray.core.alignment import broadcast
 from xarray.tests import (
     has_dask,
     raise_if_dask_computes,
